fix(inputValue): warn about invalid inputValue
