+++ test/qunit/tests.js
@@ -790,25 +790,3 @@ QUnit.test('Custom content', (assert) => {
     done()
   })
 })
-
-QUnit.test('inputValue as a Promise', (assert) => {
-  const inputTypes = ['text', 'email', 'number', 'tel', 'textarea']
-  const done = assert.async(inputTypes.length)
-  const value = '1.1 input value'
-  const inputValue = new Promise((resolve) => {
-    resolve('1.1 input value')
-  })
-  inputTypes.forEach(input => {
-    SwalWithoutAnimation({
-      input,
-      inputValue,
-      onOpen: (modal) => {
-        setTimeout(() => {
-          const inputEl = input === 'textarea' ? modal.querySelector('.swal2-textarea') : modal.querySelector('.swal2-input')
-          assert.equal(inputEl.value, input === 'number' ? parseFloat(value) : value)
-          done()
-        }, TIMEOUT)
-      }
-    })
-  })
-})
